Extract flattening info to a func, bugfix in converting map to slice, minor improvements
